requirements.txt
firecrawl-py>=0.0.16
python-dotenv>=1.0.0
